Extract widget building into own class.
